requirements.txt
flask
pandas
scikit-learn
tensorflow
pillow
numpy
joblib
requests